ila: First cut at ilad and ilactld daemons
ilad is the ILA daemon. This is run on ILA nodes. It monitors the
ILA mapping database and populates ILA routes in the kernel per
the mapping database.

ilactld is the ILA control daemon. This implements the control plane
of ILA routers. It monitors the ILA identifier database and the
locator database. It writes entries to the ILA mapping database
which are then read by the ILA daemon.