ilac: Command line interface to ILA databases
ilac (ila.py) provides the command line interface to manipulate
the ILA databases (the mapping, identifier, and locator database).
This includes commands to list the database, get specific entries,
flush the database, write entries, and destroy entries. The
identifier database includes the attach and unattach commands to
attach an identifier to a locator and unattach an identifier from
a locator.