ila: Redis configuration files
Configuration files for Redis for ILA. Three instances of Redis are
run. The configuration files are named by the ports that the instance
of Redis listens on.

Port 6379 is the default port for the mapping database.
Port 6380 is the default port for the identifier database.
Port 6381 is the default port for the locatorr database.