test: Add configuration scripts to test ILA-N and ILA-R
test_setup-N configures an ILA-N with a locator address.
'ilactl ILA-N start' is called to start the ILA-N node.

test_setup-R configured an ILA-R and ILA-M. Arguments include
a single locator to set in the locator database.

Both commands allow a list of SIR addresses to be configured
for testing connectivity.